+++ cs/test/comm/Epoxy/EpoxyListenerTests.cs
@@ -4,7 +4,6 @@
 namespace UnitTest.Epoxy
 {
     using System;
-    using System.Diagnostics;
     using System.Net;
     using System.Net.Sockets;
     using System.Threading;
@@ -28,7 +27,7 @@ namespace UnitTest.Epoxy
         public async Task ListenOnPortZero_ActuallyListensOnSomeOtherPort()
         {
             EpoxyTransport transport = MakeTransport();
-            listener = transport.MakeListener(localhostEndpoint);
+            var listener = transport.MakeListener(localhostEndpoint);
 
             await listener.StartAsync();
 
@@ -39,7 +38,7 @@ namespace UnitTest.Epoxy
         public async Task ConnectedEvent_HasRightRemoteEndpointDetails()
         {
             EpoxyTransport transport = MakeTransport();
-            listener = transport.MakeListener(localhostEndpoint);
+            var listener = transport.MakeListener(localhostEndpoint);
 
             EpoxyConnection remoteConnection = null;
             var connectedEventDone = new ManualResetEventSlim(initialState: false);
@@ -58,8 +57,6 @@ namespace UnitTest.Epoxy
 
             Assert.AreEqual(connection.LocalEndPoint, remoteConnection.RemoteEndPoint);
             Assert.AreEqual(connection.RemoteEndPoint, remoteConnection.LocalEndPoint);
-
-            await transport.StopAsync();
         }
 
         [Test]
@@ -70,7 +67,7 @@ namespace UnitTest.Epoxy
             const string DisconnectMessage = "Go away!";
 
             EpoxyTransport transport = MakeTransport();
-            listener = transport.MakeListener(localhostEndpoint);
+            var listener = transport.MakeListener(localhostEndpoint);
 
             var connectedEventDone = new ManualResetEventSlim(initialState: false);
             listener.Connected += (sender, args) =>
@@ -107,7 +104,7 @@ namespace UnitTest.Epoxy
         public async Task DisconnectedEvent_ClientDisconnects_GetsFired()
         {
             EpoxyTransport transport = MakeTransport();
-            listener = transport.MakeListener(localhostEndpoint);
+            var listener = transport.MakeListener(localhostEndpoint);
 
             var disconnectedEventDone = new ManualResetEventSlim(initialState: false);
             EpoxyConnection disconnectedConnection = null;
@@ -132,7 +129,7 @@ namespace UnitTest.Epoxy
         public async Task OneConnectionStalledDuringHandshake_CanAcceptAnother()
         {
             EpoxyTransport transport = MakeTransport();
-            listener = transport.MakeListener(localhostEndpoint);
+            var listener = transport.MakeListener(localhostEndpoint);
             await listener.StartAsync();
 
             var noHandshakeConnection = new TcpClient();
@@ -143,15 +140,13 @@ namespace UnitTest.Epoxy
             var connectTask = transport.ConnectToAsync(localhostAddress);
             bool didConnect = connectTask.Wait(TimeSpan.FromSeconds(10));
             Assert.IsTrue(didConnect, "Timed out waiting for connection to be established.");
-
-            await transport.StopAsync();
         }
 
         [Test]
         public async Task StopAsync_ClosesAllOutstandingConnections()
         {
             EpoxyTransport transport = MakeTransport();
-            listener = transport.MakeListener(localhostEndpoint);
+            var listener = transport.MakeListener(localhostEndpoint);
 
             await listener.StartAsync();
             var clientConnection = await transport.ConnectToAsync(localhostAddress);
@@ -163,9 +158,10 @@ namespace UnitTest.Epoxy
                     AnyServiceName, AnyMethodName, AnyMessage, CancellationToken.None));
         }
 
-        private static EpoxyTransport MakeTransport()
+        private EpoxyTransport MakeTransport()
         {
             var transport = new EpoxyTransportBuilder().Construct();
+            transports.Add(transport);
             return transport;
         }
     }